feat: Add useful ASG group metrics (`TOTAL_INSTANCES`, etc) by default
When diagnosing performance issues with EC2 apps, it's really useful to know the historic context of how many EC2 instances have been present in the Auto-Scaling Group - like, [_"how long have we been running MAPI mobile-fronts with 29 instances?"_](https://github.com/guardian/mobile-apps-api/pull/2865#issuecomment-2000014076) - unfortunately, by default, ASGs don't record these metrics and they need to be explicitly enabled.

There's a cost associated with every CloudWatch metric recorded, and there's always a pressure for costs to be kept down. But it's not really possible to know in advance when we're going to need these metrics, and when we need them, we really do need them - otherwise performance diagnosis is missing crucial parts of the puzzle.

To minimise the cost here, we're only adding `TOTAL_INSTANCES` & `IN_SERVICE_INSTANCES` here, rather than [all 8 ASG group metrics](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_autoscaling.GroupMetric.html#properties).

## Cost per ASG with 2 metrics enabled

https://calculator.aws/#/estimate?id=c593ec622774e2167c447f8d797ab4f04f9226be

* per month: $1.48
* per year: $18

There are two costs associated with any ASG metric being stored:

* The cost of the metric itself: $0.30 per month
* The monthly cost of the `PutMetricData` call made 43800 times ([once a minute](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_autoscaling.CfnAutoScalingGroup.MetricsCollectionProperty.html#granularity)): $0.44

## Cost on example AWS account

The `media-service` account is probably above-average in terms of the number of ASGs it has: **40**. If all ASGs in the account had the 2 metrics enabled, this would have a total cost for the account of:

* per month: $59.20 ($1.48 * 40)

The current monthly spend in the `media-service` account is about $16000, with $360 being Cloudwatch spend. From this, my belief is that an additional $59.20 in Cloudwatch spend for these metrics is not excessive.