feat(experimental-ec2-pattern): Add `buildIdentifier` prop
A requirement of using `GuEc2AppExperimental` is the filename of the application artifact should include the build identifier.

Add a required `buildIdentifier` prop to `GuEc2AppExperimental` in an attempt to use the type-system to enforce the requirement.

Tagging the launch template also means launched instances receive the tag, and added as a log line marker in Central ELK if using devx-logs to ship logs.