feat(experimental-ec2-pattern): Echo `RiffRaffDeploymentId` in user-data
Today, a redeployment of the same build is a no-op as there is no CloudFormation changeset.
This means, for example, scheduled deployments are not possible.
If we want to redeploy `main` we have to trigger a new build within GitHub.
This is quite a slow process.

This change adds a new CloudFormation parameter, `RiffRaffDeploymentId`,
to be set by Riff-Raff during deployment (see https://github.com/guardian/riff-raff/pull/1469).
This parameter is echoed out in the user-data. This means a deployment of the same build
creates a CloudFormation changeset with a new launch template and consequently the running EC2 instances are cycled.
This enables scheduled deployments and also means the run book for services using this experimental pattern become simpler; this from [MAPI's run book](https://docs.google.com/document/d/10esK5sdFKaPsq8WwR94OA-FfAU6riMs4LsEqPUGDtr0/edit?tab=t.0) goes away:

> **For mobile-sport**
> Rerun the [relevant CI build](https://github.com/guardian/mobile-apps-api/actions/workflows/mobile-sport.yml) (using the Run workflow button) for the last main commit of the [MAPI repo](https://github.com/guardian/mobile-apps-api). Once the build completes, Riff-Raff should automatically deploy it.

Co-authored-by: Jacob Winch <jacobwinch@users.noreply.github.com>